Big UPDATE!
* completely reworked the ComponentHandler
* implemented much more multi-threading where it made sense
* added "Gaming Mode" for real-time graphic drawing..
* added real-time execution for single operations on components
* added a customizable delay (default = 250ms) for interactions on
components

The main goal of this version was it to reduce the CPU usage.
In average the usage is now below 60% instead of 95%.
In "Gaming Mode" this usage will of course stay the same again.
I will try to lower the usage even more if possible before I add
new features..