Improved MouseAdapter by using the event sources rather than MouseInfo
Also this fix was necessary in order now to have the cursor location
relative to the Viewport.
Before the cursor location was only "guessed" and measured by eye what
it would be about.
That was because on JFrames (which a Display is too) there were the
boundaries / outer corners included within the location too.
Hence, on Linux Ubuntu the boundaries were differntly thick than the
ones from Windows, so I decided that needs to be fixed gently now..