Re-added rectangles with the class GRectangle with easier features
- also added the GStyle variable "focusable":
You can enable this boolean type variable in order to
tell the mouse recognition system to ignore components
with this option enabled.
This will especially effect the priority recognition of all GLayers,
meaning if you have a button placed behind a rectangle (for example),
the button behind is recognized instead (if you have "focusable" turned
off for the rectangle).