Improved further readibility and added GValueManager:
- the value manager is used internally by the components which require
an input (of type String) or a text to be displayed
- e.g, the value manager is used by GTextfield or GDescription as well
as GButton and GPolyButton

Note!

It turned out there is one bug now:
If you simulate a GRectangle in a GWindow it can block components behind
it though the rectangle should be clipped actually (try the sample
application and move your cursor on the button "MOVE TO THE RIGHT"; it
will only trigger when hovering the lower part of it..).